fix: Externalized HTML templates, and splitted multi components file into single
