+++ ui/src/app/lib/atlasmap-data-mapper/components/collapsable-header.component.html
@@ -1,0 +1,4 @@
+<div class="CollapsableCardHeader" (click)="handleMouseClick($event)">
+  <h2 class="card-pf-title">
+    <i [attr.class]="getCSSClass()"></i>{{ title }}</h2>
+</div>
