+++ ui/src/app/lib/atlasmap-data-mapper/components/data-mapper-example-host.component.ts
@@ -26,9 +26,7 @@ import { DataMapperAppComponent } from './data-mapper-app.component';
 
 @Component({
   selector: 'data-mapper-example-host',
-  template: `
-        <data-mapper #dataMapperComponent></data-mapper>
-    `,
+  template: '<data-mapper #dataMapperComponent></data-mapper>',
   providers: [MappingManagementService, ErrorHandlerService, DocumentManagementService],
 })
