+++ ui/src/app/lib/atlasmap-data-mapper/components/mapping/lookup-table.component.html
@@ -1,0 +1,9 @@
+<div class="LookupTableComponent" *ngIf="data">
+  <div class="lookupTableRow" *ngFor="let d of data">
+    <label>{{ d.sourceEnumValue }}</label>
+    <select #outputSelect [ngModel]="d.selectedTargetEnumValue" [attr.sourceValue]="d.sourceEnumValue">
+      <option *ngFor="let targetEnumValue of d.targetEnumValues" [ngValue]="targetEnumValue" [attr.enumvalue]="targetEnumValue">{{ targetEnumValue }}
+      </option>
+    </select>
+  </div>
+</div>
