+++ ui/src/app/lib/atlasmap-data-mapper/components/mapping/mapping-list-field.component.html
@@ -1,0 +1,16 @@
+<ng-template #tolTemplate>
+  <div class="fieldDetailTooltip" *ngIf="displayParentObject()">
+    <label class="parentObjectName">
+      <i [attr.class]="isSource ? 'fa fa-hdd-o' : 'fa fa-download'"></i>
+      {{ getParentObjectName() }}
+    </label>
+    <label>{{ getFieldPath() }}</label>
+    <label *ngIf="displayParentObject() && mappedField.field.type">({{ mappedField.field.type }})</label>
+    <div class="clear"></div>
+  </div>
+</ng-template>
+
+<label class="fieldPath" [tooltip]="tolTemplate" placement="bottom" [isDisabled]="!displayParentObject()">
+  {{ getFieldPath() }}
+  <i class="fa fa-bolt" *ngIf="mappedField.actions.length != 0"></i>
+</label>
