+++ ui/src/app/lib/atlasmap-data-mapper/components/mapping/mapping-list.component.ts
@@ -22,129 +22,8 @@ import { MappingModel, FieldMappingPair, MappedField } from '../../models/mappin
 import { DocumentDefinition } from '../../models/document-definition.model';
 
 @Component({
-  selector: 'mapping-list-field',
-  template: `
-        <ng-template #tolTemplate>
-            <div class="fieldDetailTooltip" *ngIf="displayParentObject()">
-                <label class="parentObjectName">
-                    <i [attr.class]="isSource ? 'fa fa-hdd-o' : 'fa fa-download'"></i>
-                    {{ getParentObjectName() }}
-                </label>
-                <label>{{ getFieldPath() }}</label>
-                <label *ngIf="displayParentObject() && mappedField.field.type">({{ mappedField.field.type }})</label>
-                <div class="clear"></div>
-            </div>
-        </ng-template>
-
-        <label class="fieldPath" [tooltip]="tolTemplate" placement="bottom" [isDisabled]="!displayParentObject()">
-            {{ getFieldPath() }}
-            <i class="fa fa-bolt" *ngIf="mappedField.actions.length != 0"></i>
-        </label>
-    `,
-})
-
-export class MappingListFieldComponent {
-  @Input() mappedField: MappedField;
-  @Input() isSource: boolean;
-  @Input() cfg: ConfigModel;
-
-  getSourceIconCSSClass(): string {
-    return this.isSource ? 'fa fa-hdd-o' : 'fa fa-download';
-  }
-
-  getFieldPath(): string {
-    if (this.mappedField == null || this.mappedField.field == null
-      || (this.mappedField.field == DocumentDefinition.getNoneField())) {
-      return '[None]';
-    }
-    return this.mappedField.field.getFieldLabel(ConfigModel.getConfig().showTypes, true);
-  }
-
-  displayParentObject(): boolean {
-    if (this.mappedField == null || this.mappedField.field == null
-      || this.mappedField.field.docDef == null
-      || (this.mappedField.field == DocumentDefinition.getNoneField())) {
-      return false;
-    }
-    return true;
-  }
-
-  getParentObjectName() {
-    if (this.mappedField == null || this.mappedField.field == null || this.mappedField.field.docDef == null) {
-      return '';
-    }
-    return this.mappedField.field.docDef.getName(ConfigModel.getConfig().showTypes);
-  }
-}
-
-@Component({
   selector: 'mapping-list',
-  template: `
-        <div class="dataMapperItemList mappingList">
-            <div class="card-pf">
-                <div class="card-pf-heading">
-                    <h2 class="card-pf-title">
-                        <div class="name">
-                        <i class="fa fa-table"></i>
-                            <label>Mappings</label>
-                        </div>
-                        <i (click)="toggleSearch()" [attr.class]="getSearchIconCSSClass()"></i>
-                        <div class="clear"></div>
-                    </h2>
-                    <div class="searchHeaderWrapper">
-                        <div *ngIf="searchMode" class="searchBox">
-                            <input type="text" #searchFilterBox id="search-filter-box" [(ngModel)]="searchFilter"
-                                (keyup)="search(searchFilterBox.value)" placeholder="Search" [focus]="true" />
-                            <i class="fa fa-close searchBoxCloseIcon link" (click)="toggleSearch()"></i>
-                            <div class="clear"></div>
-                        </div>
-                        <div [attr.class]="getRowTitleCSSClass()">
-                            <label class="sources"><i class="fa fa-hdd-o"></i>Sources</label>
-                            <label class="targets"><i class="fa fa-download"></i>Targets</label>
-                            <label class="type"><i class="fa fa-sliders"></i>Type</label>
-                            <div class="clear"></div>
-                        </div>
-                        <div class="clear"></div>
-                    </div>
-                </div>
-                <div [attr.class]="getItemsCSSClass()">
-                    <div [attr.class]="getMappingRowsCSSClass()">
-                        <div *ngFor="let mapping of getMappings(); let index=index;"
-                            [attr.class]="getMappingCSSClass(mapping, index)" (click)="selectMapping(mapping)">
-                            <div *ngFor="let fieldPair of mapping.fieldMappings" class="itemRow">
-                                <div class="sourceFieldNames fieldNames">
-                                    <mapping-list-field *ngFor="let mappedField of getMappedFields(fieldPair, true)"
-                                        [mappedField]="mappedField" [isSource]="true" [cfg]="cfg"></mapping-list-field>
-                                    <div class="clear"></div>
-                                </div>
-                                <div class="targetFieldNames fieldNames">
-                                    <mapping-list-field *ngFor="let mappedField of getMappedFields(fieldPair, false)"
-                                        [mappedField]="mappedField" [isSource]="false" [cfg]="cfg"></mapping-list-field>
-                                    <div class="clear"></div>
-                                </div>
-                                <div class="transition">
-                                    <label>{{ fieldPair.transition.getPrettyName() }}</label>
-                                    <div class="clear"></div>
-                                </div>
-                                <div class="error">
-                                    <span class="pficon"
-                                          [ngClass]="mapping.validationErrors | toErrorIconClass"
-                                          *ngIf="mapping.validationErrors.length"></span>
-                                </div>
-                                <div class="clear"></div>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="noSearchResults" *ngIf="searchResultsVisible()">
-                        <label>No search results.</label>
-                        <div class="clear"></div>
-                    </div>
-                </div>
-                <div class="card-pf-heading itemCount">{{ cfg.mappings.mappings.length }} mappings</div>
-                <div class="clear"></div>
-            </div>
-        </div>
-    `,
+  templateUrl: './mapping-list.component.html',
 })
 
 export class MappingListComponent {
