+++ ui/src/app/lib/atlasmap-data-mapper/components/mapping/mapping-selection.component.ts
@@ -14,122 +14,18 @@
     limitations under the License.
 */
 
-import { Component, Input, ViewChildren, QueryList } from '@angular/core';
+import { Component, ViewChildren, QueryList } from '@angular/core';
 
-import { MappingModel, FieldMappingPair } from '../../models/mapping.model';
+import { MappingModel } from '../../models/mapping.model';
 import { ConfigModel } from '../../models/config.model';
 import { Field } from '../../models/field.model';
 
 import { ModalWindowComponent } from '../modal-window.component';
-
-@Component({
-  selector: 'mapping-selection-section',
-  template: `
-        <div [attr.class]="getClass()" (click)="handleMouseClick($event)">
-            <div class="numberWrapper"><div class="number">{{ outputNumber + 1 }}</div></div>
-            <div class="pathContainer">
-                <div class="fieldPair" *ngFor="let fieldPair of mapping.fieldMappings">
-                    <div class="sourceTargetSection sourcePaths">
-                        <label>{{ getSourceTargetLabelText(true, fieldPair) }}</label>
-                        <div class="paths" *ngFor="let sourceField of fieldPair.sourceFields">
-                            <div class="path">
-                                <div class="pathName">
-                                    <i class="fa fa-hdd-o" aria-hidden="true"></i>
-                                    {{ getFormattedOutputPath(sourceField.field.path, false) }}
-                                </div>
-                                <div class="fieldName">{{ getFormattedOutputPath(sourceField.field.path, true) }}</div>
-                                <div class="clear"></div>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="sourceTargetSection targetPaths">
-                        <label>{{ getSourceTargetLabelText(false, fieldPair) }}</label>
-                        <div class="paths" *ngFor="let targetField of fieldPair.targetFields">
-                            <div class="path">
-                                <div class="pathName">
-                                    <i class="fa fa-download" aria-hidden="true"></i>
-                                    {{ getFormattedOutputPath(targetField.field.path, false) }}
-                                </div>
-                                <div class="fieldName">{{ getFormattedOutputPath(targetField.field.path, true) }}</div>
-                                <div class="clear"></div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="clear"></div>
-        </div>
-    `,
-})
-
-export class MappingSelectionSectionComponent {
-  @Input() outputNumber: number;
-  @Input() mapping: MappingModel;
-  @Input() selectedCallback: Function;
-  @Input() selected = false;
-  @Input() selectedFieldIsSource = false;
-  @Input() parentComponent: Component;
-  @Input() isOddRow = false;
-
-  getClass(): string {
-    let cssClass = 'MappingSelectionSection';
-    if (this.selected) {
-      cssClass += ' SelectedMappingSelectionSection';
-    }
-    if (this.isOddRow) {
-      cssClass += ' odd';
-    }
-    return cssClass;
-  }
-
-  getSourceTargetLabelText(isSource: boolean, fieldPair: FieldMappingPair): string {
-    if (isSource) {
-      return (fieldPair.sourceFields.length > 0) ? 'Sources' : 'Source';
-    }
-    return (fieldPair.targetFields.length > 0) ? 'Targets' : 'Target';
-  }
-
-  getFormattedOutputPath(path: string, nameOnly: boolean) {
-    if (path == null) {
-      return '';
-    }
-    path = path.replace('.', '/');
-    const index: number = path.lastIndexOf('/');
-    const fieldName: string = (index == -1) ? path : path.substr(path.lastIndexOf('/') + 1);
-    path = (index == -1) ? '' : path.substr(0, path.lastIndexOf('/') + 1);
-    return nameOnly ? fieldName : path;
-  }
-
-  handleMouseClick(event: MouseEvent) {
-    this.selectedCallback(this);
-  }
-}
+import { MappingSelectionSectionComponent } from './mapping-selection-section.component';
 
 @Component({
   selector: 'mapping-selection',
-  template: `
-        <div class="MappingSelectionComponent" *ngIf="mappings">
-            <div class="header">
-                <div class="sourceTargetHeader">{{ selectedField.isSource() ? 'Source' : 'Target' }}</div>
-                <div class="pathHeader">
-                    <div class="pathName">{{ getFormattedOutputPath(selectedField.path, false) }}</div>
-                    <div class="fieldName">{{ getFormattedOutputPath(selectedField.path, true) }}</div>
-                    <div class="clear"></div>
-                </div>
-                <div class="clear"></div>
-                <button class="btn btn-primary addButton" (click)="addMapping()">
-                    <i class="fa fa-plus"></i>Add New Mapping
-                </button>
-            </div>
-            <div class="MappingSelectionOptions">
-                <mapping-selection-section *ngFor="let mapping of mappings; let i = index; let odd=odd; let even=even;"
-                    [mapping]="mapping" [outputNumber]="i" [selected]="i == 0" [selectedCallback]="selectionChanged"
-                    [selectedFieldIsSource]="selectedField.isSource()" [parentComponent]="this"
-                    [isOddRow]="odd" #mappingSection>
-                </mapping-selection-section>
-            </div>
-        </div>
-    `,
+  templateUrl: './mapping-selection.component.html',
 })
 
 export class MappingSelectionComponent {
