+++ ui/src/app/lib/atlasmap-data-mapper/components/namespace-edit.component.html
@@ -1,0 +1,21 @@
+<div class="PropertyEditFieldComponent">
+  <div class="form-group">
+    <label>Alias</label>
+    <input type="text" [(ngModel)]="namespace.alias" disabled="{{namespace.isTarget || !namespace.createdByUser}}">
+  </div>
+  <div class="form-group">
+    <label>URI</label>
+    <input type="text" [(ngModel)]="namespace.uri" />
+  </div>
+  <div class="form-group">
+    <label>Location URI</label>
+    <input type="text" [(ngModel)]="namespace.locationUri" />
+  </div>
+  <div class="form-group">
+    <label>Type</label>
+    <input type="checkbox" [ngModel]="namespace.isTarget" style="width:20px; vertical-align:middle;" disabled="{{!targetEnabled}}"
+      (click)="targetToggled()" />
+    <label [attr.class]="(targetEnabled ? '' : 'disabled')" style="width:105px; ">Target Namespace</label>
+    <div class="clear"></div>
+  </div>
+</div>
