+++ ui/src/app/lib/atlasmap-data-mapper/components/property-field-edit.component.ts
@@ -22,45 +22,7 @@ import { ModalWindowValidator } from './modal-window.component';
 
 @Component({
   selector: 'property-field-edit',
-  template: `
-        <div class="DataMapperEditComponent">
-            <div class="form-group">
-                <label>Name:</label>
-                <input name="name" type="text" [(ngModel)]="field.name"/>
-            </div>
-            <div class="form-group">
-                <label>Value:</label>
-                <input name="value" type="text" [(ngModel)]="field.value"/>
-            </div>
-            <div class="form-group">
-                <label>Value Type</label>
-                <select (change)="valueTypeSelectionChanged($event);" [ngModel]="valueType">
-                    <option value="BOOLEAN">Boolean</option>
-                    <option value="BYTE">Byte</option>
-                    <option value="BYTE_ARRAY">ByteArray</option>
-                    <option value="CHAR">Char</option>
-                    <option value="COMPLEX">Complex</option>
-                    <option value="DECIMAL">Decimal</option>
-                    <option value="DOUBLE">Double</option>
-                    <option value="FLOAT">Float</option>
-                    <option value="INTEGER">Integer</option>
-                    <option value="LONG">Long</option>
-                    <option value="SHORT">Short</option>
-                    <option value="STRING">String</option>
-                    <option value="TIME">Time</option>
-                    <option value="DATE">Date</option>
-                    <option value="DATE_TIME">DateTime</option>
-                    <option value="DATE_TZ">DateTZ</option>
-                    <option value="TIME_TZ">TimeTZ</option>
-                    <option value="DATE_TIME_TZ">DateTimeTZ</option>
-                    <option value="UNSIGNED_BYTE">Unsigned Byte</option>
-                    <option value="UNSIGNED_INTEGER">Unsigned Integer</option>
-                    <option value="UNSIGNED_LONG">Unsigned Long</option>
-                    <option value="UNSIGNED_SHORT">Unsigned Short</option>
-                </select>
-            </div>
-        </div>
-    `,
+  templateUrl: './property-field-edit.component.html',
 })
 
 export class PropertyFieldEditComponent implements ModalWindowValidator {
