+++ ui/src/app/lib/atlasmap-data-mapper/components/template-edit.component.html
@@ -1,0 +1,5 @@
+<div class="DataMapperEditComponent">
+  <div class="form-group">
+    <textarea [(ngModel)]="templateText" rows="16" cols="100"></textarea>
+  </div>
+</div>
